systemd: re-enable mount propagation for udevd
MountFlags's default value is shared in systemd-udevd.service. But upstream
sets MountFlags with slave just for keeping mounts done by udev private to
udevd, which causes block device mounted by udev unvisable but being busy for
host. So we revert it to shared to be propagated to host.

Original Patch:
https://patchwork.openembedded.org/patch/148237/

Change-Id: I7ff4c65debc15836ea91888d76a99a5ccbdca570
Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>